+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 0.1.4
+- Fix: `DebugConfigurationProvider` should set a default only for required values for the specific configuration type. 
+
 ## 0.1.3
 - Implement `skip_path` support based on a custom DAP request.
   - Adds support for these paths in `attach` configurations.